backend/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Pinecone Vector Database (Free Tier: 1 index, 100K vectors)
pinecone-client>=3.0.0

# LLM & Embeddings
groq>=0.4.0
sentence-transformers>=2.2.0

# HTTP & Web
httpx>=0.24.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.9
tqdm>=4.66.0
